+++ apps/web/src/services/workspace-home/store/workspace-home-page-store.ts
@@ -1,10 +1,13 @@
 import { computed, reactive } from 'vue';
 
+import dayjs from 'dayjs';
 import { defineStore } from 'pinia';
 
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
+import { getCancellableFetcher } from '@cloudforet/core-lib/space-connector/cancallable-fetcher';
 import { ApiQueryHelper } from '@cloudforet/core-lib/space-connector/helper';
 
+import type { AnalyzeResponse } from '@/schema/_common/api-verbs/analyze';
 import type { ListResponse } from '@/schema/_common/api-verbs/list';
 import type { UserConfigListParameters } from '@/schema/config/user-config/api-verbs/list';
 import type { UserConfigModel } from '@/schema/config/user-config/model';
@@ -15,9 +18,11 @@ import type { AppListParameters } from '@/schema/identity/app/api-verbs/list';
 import type { AppModel } from '@/schema/identity/app/model';
 import type { WorkspaceUserListParameters } from '@/schema/identity/workspace-user/api-verbs/list';
 import type { WorkspaceUserModel } from '@/schema/identity/workspace-user/model';
+import type { MetricDataAnalyzeParameters } from '@/schema/inventory/metric-data/api-verbs/analyze';
 import { store } from '@/store';
 
 import { useUserWorkspaceStore } from '@/store/app-context/workspace/user-workspace-store';
+import { useAllReferenceStore } from '@/store/reference/all-reference-store';
 
 import { MENU_ID } from '@/lib/menu/config';
 
@@ -26,9 +31,18 @@ import type { FavoriteItem } from '@/common/modules/favorites/favorite-button/ty
 import { FAVORITE_TYPE } from '@/common/modules/favorites/favorite-button/type';
 import { RECENT_TYPE } from '@/common/modules/navigations/type';
 
+import type { MetricDataAnalyzeResult } from '@/services/asset-inventory/types/asset-analysis-type';
+import { convertFormKeys } from '@/services/workspace-home/composables/use-workspace-home';
+import type {
+    ProviderReferenceDataMap, ProviderResourceDataItem, DailyUpdatesListItem,
+} from '@/services/workspace-home/types/workspace-home-type';
+
+
 export const useWorkspaceHomePageStore = defineStore('page-workspace-home', () => {
     const userWorkspaceStore = useUserWorkspaceStore();
     const userWorkspaceStoreGetters = userWorkspaceStore.getters;
+    const allReferenceStore = useAllReferenceStore();
+    const allReferenceGetters = allReferenceStore.getters;
 
     const state = reactive({
         recentList: [] as UserConfigModel[],
@@ -37,11 +51,18 @@ export const useWorkspaceHomePageStore = defineStore('page-workspace-home', () =
         appsTotalCount: undefined as number|undefined,
         costReportConfig: null as CostReportConfigModel|null|undefined,
         dataSource: [] as CostDataSourceModel[],
+        providers: [] as ProviderResourceDataItem[],
+        dailyUpdatesListItems: {
+            created: [],
+            deleted: [],
+        } as DailyUpdatesListItem,
     });
 
     const _getters = reactive({
+        timezone: computed(() => store.state.user.timezone),
         userId: computed<string>(() => store.state.user.userId),
         currentWorkspaceId: computed<string|undefined>(() => userWorkspaceStoreGetters.currentWorkspaceId),
+        providerMap: computed<ProviderReferenceDataMap>(() => allReferenceGetters.provider),
     });
 
     const recentListApiQuery = new ApiQueryHelper().setSort('updated_at', true);
@@ -146,6 +167,84 @@ export const useWorkspaceHomePageStore = defineStore('page-workspace-home', () =
             } catch (e) {
                 ErrorHandler.handleError(e);
                 state.dataSource = [];
+            }
+        },
+        fetchCloudServiceResources: async () => {
+            const labels = ['Server', 'Database', 'Storage'];
+
+            try {
+                await Promise.all(labels.map(async (label) => {
+                    const metricId = `metric-managed-${label.toLowerCase()}-${label !== 'Storage' ? 'count' : 'size'}`;
+                    const fetcher = getCancellableFetcher<MetricDataAnalyzeParameters, AnalyzeResponse<MetricDataAnalyzeResult>>(SpaceConnector.clientV2.inventory.metricData.analyze);
+                    const { status, response } = await fetcher({
+                        metric_id: metricId,
+                        query: {
+                            granularity: 'DAILY',
+                            group_by: ['labels.Provider'],
+                            start: dayjs.tz(dayjs.utc(), _getters.timezone).subtract(5, 'days').format('YYYY-MM-DD'),
+                            end: dayjs.tz(dayjs.utc(), _getters.timezone).format('YYYY-MM-DD'),
+                            fields: {
+                                count: {
+                                    key: 'value',
+                                    operator: 'sum',
+                                },
+                            },
+                            sort: [{ key: '_total_count', desc: true }],
+                            field_group: ['date'],
+                        },
+                    });
+
+                    if (status === 'succeed') {
+                        (response?.results || []).forEach((i) => {
+                            _getters.providerMap[i.Provider][label.toLowerCase()] = i._total_count;
+                        });
+                    }
+                }));
+                state.providers = Object.keys(_getters.providerMap).map((key) => _getters.providerMap[key]);
+            } catch (e) {
+                ErrorHandler.handleError(e);
+                state.providers = [];
+            }
+        },
+        fetchDailyUpdatesList: async (): Promise<void> => {
+            const labels = ['created', 'deleted'];
+
+            try {
+                await Promise.all(labels.map(async (label) => {
+                    const metricId = `metric-managed-${label.toLowerCase()}-count`;
+                    const fetcher = getCancellableFetcher<MetricDataAnalyzeParameters, AnalyzeResponse<MetricDataAnalyzeResult>>(SpaceConnector.clientV2.inventory.metricData.analyze);
+                    const { status, response } = await fetcher({
+                        metric_id: metricId,
+                        query: {
+                            granularity: 'DAILY',
+                            group_by: ['labels.Provider', 'labels.Cloud Service Group', 'labels.Cloud Service Type'],
+                            start: dayjs.tz(dayjs.utc(), _getters.timezone).format('YYYY-MM-DD'),
+                            end: dayjs.tz(dayjs.utc(), _getters.timezone).format('YYYY-MM-DD'),
+                            fields: {
+                                count: {
+                                    key: 'value',
+                                    operator: 'sum',
+                                },
+                            },
+                            sort: [{ key: '_total_count', desc: true }],
+                            field_group: ['date'],
+                        },
+                    });
+
+                    if (status === 'succeed') {
+                        const results = convertFormKeys(response.results || []);
+                        state.dailyUpdatesListItems[label] = results.map((i) => ({
+                            cloud_service_group: i.cloud_service_group,
+                            cloud_service_type: i.cloud_service_type,
+                            total_count: i._total_count,
+                            provider: i.provider,
+                            [`${label}_count`]: i.count[0].value,
+                        }));
+                    }
+                }));
+            } catch (e) {
+                ErrorHandler.handleError(e);
+                state.dailyUpdatesListItems = { created: state.dailyUpdatesListItems.created || [], deleted: state.dailyUpdatesListItems.deleted || [] };
             }
         },
     };
